requirements.txt
alembic==1.16.2
asgiref==3.8.1
attrs==25.3.0
cascade==0.1
coverage==7.9.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
greenlet==3.2.3
imaplib2==3.6
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
progressbar==2.5
Pygments==2.19.1
PyJWT==2.9.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
